+++ EasyMusicPlayer/SearchViewState.swift
@@ -7,9 +7,12 @@ protocol SearchViewStating {
     var numberOfSections: Int { get }
     var isEmptyLabelHidden: Bool { get }
     var emptyText: String { get }
+    var isLoading: Bool { get }
 
     func item(at indexPath: IndexPath) -> Track
     func cellViewState(at indexPath: IndexPath) -> SearchCellViewStating
+
+    func copy(isLoading: Bool) -> SearchViewState
 }
 
 struct SearchViewState: SearchViewStating {
@@ -22,10 +25,12 @@ struct SearchViewState: SearchViewStating {
         return !items.isEmpty
     }
     let emptyText: String = L10n.searchViewEmptyText
+    let isLoading: Bool
     private let items: [MPMediaItem]
 
-    init(items: [MPMediaItem]) {
+    init(items: [MPMediaItem], isLoading: Bool) {
         self.items = items
+        self.isLoading = isLoading
     }
 
     func item(at indexPath: IndexPath) -> Track {
@@ -36,5 +41,11 @@ struct SearchViewState: SearchViewStating {
         let item = Track(mediaItem: items[indexPath.row], artworkSize: CGSize(width: 100, height: 100))
         let title = "\(item.artist)\n\(item.title)"
         return SearchCellViewState(title: title, image: item.artwork)
+    }
+}
+
+extension SearchViewState {
+    func copy(isLoading: Bool) -> SearchViewState {
+        return SearchViewState(items: items, isLoading: isLoading)
     }
 }