+++ EasyMusicPlayer/_Simulator.swift
@@ -81,19 +81,47 @@ final class DummyMediaItem: MPMediaItem {
         }
         super.init()
     }
+
+    override var hash: Int {
+        return Int(_persistentID)
+    }
 }
 
 // MARK: - DummyMediaQuery
 
 final class DummyMediaQuery: MPMediaQuery {
+    private static let library: [DummyMediaItem] = {
+        // REINSTALL APP ON SIMULATOR AFTER CHANGING THIS
+        // use to test a small, specific library
+        return [DummyMediaItem(asset: .normal, id: 0),
+                DummyMediaItem(asset: .endSilence, id: 1),
+                DummyMediaItem(asset: .normal, id: 2)]
+//        // use to test a large number of items - might be slow on load
+//        return (0..<50_000).map {
+//            DummyMediaItem(asset: .normal, artist: UUID().uuidString, title: UUID().uuidString, id: $0)
+//        }
+    }()
+    private var id: MPMediaEntityPersistentID?
+
     override var items: [MPMediaItem]? {
         if ProcessInfo.processInfo.environment["XCTestConfigurationFilePath"] == nil {
-            return [DummyMediaItem(asset: .normal, id: 0),
-                    DummyMediaItem(asset: .endSilence, id: 1),
-                    DummyMediaItem(asset: .normal, id: 2)]
+            if let id = id {
+                return [type(of: self).library[Int(id)]]
+            } else {
+                return type(of: self).library
+            }
         } else {
             return super.items
         }
+    }
+
+    override func addFilterPredicate(_ predicate: MPMediaPredicate) {
+        guard let predicate = predicate as? MPMediaPropertyPredicate else { return }
+        id = predicate.value as? MPMediaEntityPersistentID
+    }
+
+    override func removeFilterPredicate(_ predicate: MPMediaPredicate) {
+        id = nil
     }
 }
 #endif