+++ EasyMusicPlayer/_Simulator.swift
@@ -91,15 +91,15 @@ final class DummyMediaItem: MPMediaItem {
 
 final class DummyMediaQuery: MPMediaQuery {
     private static let library: [DummyMediaItem] = {
-        // REINSTALL APP ON SIMULATOR AFTER CHANGING THIS
-        // use to test a small, specific library
-        return [DummyMediaItem(asset: .normal, id: 0),
-                DummyMediaItem(asset: .endSilence, id: 1),
-                DummyMediaItem(asset: .normal, id: 2)]
+//        // REINSTALL APP ON SIMULATOR AFTER CHANGING THIS
+//        // use to test a small, specific library
+//        return [DummyMediaItem(asset: .normal, id: 0),
+//                DummyMediaItem(asset: .endSilence, id: 1),
+//                DummyMediaItem(asset: .normal, id: 2)]
 //        // use to test a large number of items - might be slow on load
-//        return (0..<50_000).map {
-//            DummyMediaItem(asset: .normal, artist: UUID().uuidString, title: UUID().uuidString, id: $0)
-//        }
+        return (0..<50_000).map {
+            DummyMediaItem(asset: .normal, artist: UUID().uuidString, title: UUID().uuidString, id: $0)
+        }
     }()
     private var id: MPMediaEntityPersistentID?